chore: Splitting tests (#3)
* chore: draft

* chore: splitting tests
